Add error checking to local store
Also fix indentation.

npm test passes

+++ src/store/local.ts
@@ -1,6 +1,7 @@
 //
 // Copyright (c) 2018 Nathan Fiedler
 //
+import fs = require('fs')
 import path = require('path')
 import fx = require('fs-extra')
 
@@ -8,28 +9,34 @@ import fx = require('fs-extra')
  * Local disk implementation of the `Store` interface.
  */
 export class LocalStore {
-    readonly basepath: string
+  readonly basepath: string
 
-    /**
-     * Construct a new instance of LocalStore.
-     *
-     * @param basepath location where pack files will be stored.
-     */
-    constructor(basepath: string) {
-        this.basepath = basepath
-    }
+  /**
+   * Construct a new instance of LocalStore.
+   *
+   * @param basepath location where pack files will be stored.
+   */
+  constructor(basepath: string) {
+    this.basepath = basepath
+  }
 
-    storePack(packfile: string, bucket: string, object: string): void {
-        const buckdir = path.join(this.basepath, bucket)
-        fx.ensureDirSync(buckdir)
-        const destfile = path.join(buckdir, object)
-        fx.moveSync(packfile, destfile)
+  storePack(packfile: string, bucket: string, object: string): void {
+    if (!fs.existsSync(packfile)) {
+      throw new Error('missing file: ' + packfile)
     }
+    const buckdir = path.join(this.basepath, bucket)
+    fx.ensureDirSync(buckdir)
+    const destfile = path.join(buckdir, object)
+    fx.moveSync(packfile, destfile)
+  }
 
-    retrievePack(bucket: string, object: string, outfile: string): void {
-        const buckdir = path.join(this.basepath, bucket)
-        fx.ensureDirSync(buckdir)
-        const packfile = path.join(buckdir, object)
-        fx.copySync(packfile, outfile)
+  retrievePack(bucket: string, object: string, outfile: string): void {
+    const buckdir = path.join(this.basepath, bucket)
+    fx.ensureDirSync(path.dirname(outfile))
+    const packfile = path.join(buckdir, object)
+    if (!fs.existsSync(packfile)) {
+      throw new Error('missing pack file: ' + packfile)
     }
+    fx.copySync(packfile, outfile)
+  }
 }
